shadow-dom/shadow-trees/test-003.html

<!DOCTYPE html>
<!--
Distributed under both the W3C Test Suite License [1] and the W3C
3-clause BSD License [2]. To contribute to a W3C Test Suite, see the
policies and contribution forms [3].

[1] http://www.w3.org/Consortium/Legal/2008/04-testsuite-license
[2] http://www.w3.org/Consortium/Legal/2008/03-bsd-license
[3] http://www.w3.org/2004/10/27-testcases
-->
<html>
<head>
<title>Shadow DOM Test: A_04_00_03</title>
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<link rel="author" title="Mikhail Fursov" href="mailto:mfursov@unipro.ru">
<link rel="help" href="http://www.w3.org/TR/2013/WD-shadow-dom-20130514/#shadow-dom-subtrees">
<meta name="assert" content="Shadow DOM subtree takes place of the shadow host's content when rendered">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script>
// check that elements from shadow tree are rendered as a part of the host document
test(unit(function (ctx) {

    var d = newRenderedHTMLDocument(ctx);
    var div = d.createElement('div');
    d.body.appendChild(div);
    assert_true(div.offsetTop > 0, 'Expected: the host element is rendered before the check. Actual: the host element is not rendered');

    var s = createSR(div);
    var hr = d.createElement('hr');
    assert_true(hr.offsetTop == 0, 'Expected: The element to be added to shadow is not rendered before the check');

    s.appendChild(hr);
    assert_true(hr.offsetTop >= div.offsetTop, 'Expected: The element is rendered after adding to shadow tree');

}), 'A_04_00_03_T01');


//check that shadow root content is replaced by the shadow tree
test(unit(function (ctx) {

    var d = newRenderedHTMLDocument(ctx);
    var div = d.createElement('div');
    div.innerHTML = '<span id="spandex">Shadow Root content to be replaced</span>';
    d.body.appendChild(div);

    assert_true(d.querySelector('#spandex').offsetTop > 0, 'Expected: The host element content is ' +
        'rendered before the check');

    var s = createSR(div);
    var hr = d.createElement('hr');
    assert_true(hr.offsetTop == 0, "Expected: The element to be added to shadow is not rendered before the check");

    s.appendChild(hr);
    assert_true(hr.offsetTop >= div.offsetTop, "Expected: The element is rendered after adding to shadow tree");
    assert_equals(d.querySelector('#spandex').offsetTop, 0, 'The shadow host element content should ' +
        'be replaced by the shadow tree');

}), 'A_04_00_03_T02');
</script>
</body>
</html>
